tests/wpt/tests/css/css-mixins/function-parameter-types.tentative.html

<!DOCTYPE html>
<title>Custom Functions: Typed parameters</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/12315">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

<div id=target></div>
<div id=main></div>

<!-- To pass, a test must produce matching computed values for --actual and
     --expected on #target. -->

<template data-name="A parameter retains its type">
  <style>
    @function --f(--c <color>) {
      result: if(
        style(--c:red): PASS;
        else: FAIL);
    }
    #target {
      --actual: --f(#f00);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="A parameter type acts as a local registration">
  <style>
    @function --f(--c <color>) {
      --c:#00f; /* <color> */
      result: if(
        style(--c:blue): PASS;
        else: FAIL);
    }
    #target {
      --actual: --f(#000);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="A parameter retains its type (parent stack frame)">
  <style>
    @function --f() {
      result: if(
        style(--c:red): PASS;
        else: FAIL);
    }
    @function --g(--c <color>) {
      result: --f();
    }
    #target {
      --actual: --g(#f00);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="A parameter type acts as a local registration (parent stack frame)">
  <style>
    @function --f() {
      result: if(
        style(--c:blue): PASS;
        else: FAIL);
    }
    @function --g(--c <color>) {
      --c:#00f; /* <color> */
      result: --f();
    }
    #target {
      --actual: --g(#f00);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="Universally typed parameter can shadow other parameters">
  <style>
    @function --f(--c type(*)) {
      result: if(
        style(--c:red): FAIL1;
        style(--c:#f00): FAIL2;
        style(--c:lime): FAIL3;
        style(--c:#0f0): PASS;
        else: FAIL4);
    }
    @function --g(--c <color>) {
      result: --f(#0f0);
    }
    #target {
      --actual: --g(#f00);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="Invalid value for typed local becomes IACVT">
  <style>
    @function --f(--c <color>) {
      --c: 3;
      result: var(--c, PASS);
    }
    #target {
      --actual: --f(#f00);
      --expected: PASS;
    }
  </style>
</template>

<template data-name="if() within @function can query registered custom property">
  <style>
    @property --c {
      syntax: "<color>";
      inherits: false;
      initial-value: black;
    }
    @function --f() {
      result: if(style(--c:#f00): PASS; else: FAIL);
    }
    #target {
      --c: red;
      --actual: --f();
      --expected: PASS;
    }
  </style>
</template>

<script>
  test_all_templates();
</script>
